# docs/swagger.yaml
basePath: /api
definitions:
  dto.AddGroupMemberRequest:
    properties:
      role:
        enum:
        - admin
        - member
        type: string
      user_id:
        type: string
    required:
    - role
    - user_id
    type: object
  dto.CreateGroupRequest:
    properties:
      created_by:
        type: string
      currency:
        type: string
      description:
        type: string
      name:
        type: string
      type:
        type: string
    required:
    - created_by
    - currency
    - name
    - type
    type: object
  dto.CreateUserRequest:
    properties:
      email:
        type: string
      firebase_uid:
        type: string
      full_name:
        type: string
      password:
        type: string
      phone:
        type: string
      username:
        type: string
    required:
    - email
    - full_name
    - username
    type: object
  dto.LeaveGroupRequest:
    properties:
      user_id:
        type: string
    required:
    - user_id
    type: object
  dto.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  dto.SignupRequest:
    properties:
      email:
        type: string
      firebase_uid:
        type: string
      full_name:
        type: string
      password:
        minLength: 6
        type: string
      phone:
        type: string
      username:
        type: string
    required:
    - email
    - firebase_uid
    - full_name
    - password
    - username
    type: object
  dto.TokenResponse:
    properties:
      email:
        type: string
      expires_in:
        type: string
      id_token:
        type: string
      refresh_token:
        type: string
    type: object
  dto.UpdateGroupRequest:
    properties:
      created_by:
        type: string
      currency:
        type: string
      description:
        type: string
      name:
        type: string
      type:
        type: string
    type: object
  dto.UpdateUserRequest:
    properties:
      currency_pref:
        type: string
      full_name:
        type: string
      password:
        type: string
      phone:
        type: string
      username:
        type: string
    type: object
  models.Group:
    properties:
      createdAt:
        type: string
      createdBy:
        type: string
      currency:
        type: string
      description:
        type: string
      id:
        type: string
      inviteCode:
        type: string
      isActive:
        type: boolean
      isPersonal:
        type: boolean
      name:
        type: string
      type:
        type: string
    type: object
  models.GroupMember:
    properties:
      groupID:
        type: string
      id:
        type: string
      joinedAt:
        type: string
      leftAt:
        type: string
      nickname:
        type: string
      role:
        type: string
      userID:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Expense splitting and group finance management backend.
  title: PRISM API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Validates credentials. Use the Firebase SDK separately to get an
        ID token.
      parameters:
      - description: Login payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: user object
          schema:
            additionalProperties: true
            type: object
        "400":
          description: validation error
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: invalid credentials
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login with email and password
      tags:
      - auth
  /auth/signup:
    post:
      consumes:
      - application/json
      description: Creates a user in the DB and stamps a Firebase custom claim with
        the DB UUID.
      parameters:
      - description: Signup payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.SignupRequest'
      produces:
      - application/json
      responses:
        "201":
          description: created user
          schema:
            additionalProperties: true
            type: object
        "400":
          description: validation error
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: email or username already in use
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new user
      tags:
      - auth
  /auth/token:
    post:
      consumes:
      - application/json
      description: Exchanges email and password for a Firebase ID token. Use the returned
        idToken as the Bearer token for all protected endpoints.
      parameters:
      - description: Email and password
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Firebase ID token and metadata
          schema:
            $ref: '#/definitions/dto.TokenResponse'
        "400":
          description: validation error
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: invalid credentials
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get Firebase ID token
      tags:
      - auth
  /groups:
    get:
      description: Returns all active groups.
      produces:
      - application/json
      responses:
        "200":
          description: list of groups
          schema:
            items:
              $ref: '#/definitions/models.Group'
            type: array
        "500":
          description: internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List all groups
      tags:
      - groups
    post:
      consumes:
      - application/json
      description: Creates a new expense-splitting group.
      parameters:
      - description: Group payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateGroupRequest'
      produces:
      - application/json
      responses:
        "201":
          description: created group
          schema:
            $ref: '#/definitions/models.Group'
        "400":
          description: validation error
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a group
      tags:
      - groups
  /groups/{id}:
    delete:
      description: Soft-deletes a group by setting is_active = false.
      parameters:
      - description: Group UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: message
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: invalid id
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a group
      tags:
      - groups
    get:
      description: Returns a single group by UUID.
      parameters:
      - description: Group UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: group
          schema:
            $ref: '#/definitions/models.Group'
        "400":
          description: invalid id
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: group not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get group by ID
      tags:
      - groups
    put:
      consumes:
      - application/json
      description: Updates mutable fields on a group — name, description, type, currency.
      parameters:
      - description: Group UUID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateGroupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: updated group
          schema:
            $ref: '#/definitions/models.Group'
        "400":
          description: invalid id or payload
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a group
      tags:
      - groups
  /groups/{id}/balances:
    get:
      description: Returns the net balance for each member in the group.
      parameters:
      - description: Group UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: balances per member
          schema:
            additionalProperties: true
            type: object
        "400":
          description: invalid id
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get group balances
      tags:
      - groups
  /groups/{id}/leave:
    post:
      consumes:
      - application/json
      description: Removes the specified user from the group. Pass the user_id in
        the request body.
      parameters:
      - description: Group UUID
        in: path
        name: id
        required: true
        type: string
      - description: User to remove
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.LeaveGroupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: message
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: invalid id or payload
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Leave a group
      tags:
      - groups
  /groups/{id}/members:
    get:
      description: Returns all current members of the specified group.
      parameters:
      - description: Group UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: members
          schema:
            items:
              $ref: '#/definitions/models.GroupMember'
            type: array
        "400":
          description: invalid id
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List group members
      tags:
      - group-members
    post:
      consumes:
      - application/json
      description: Adds a user to the specified group with a given role (admin or
        member).
      parameters:
      - description: Group UUID
        in: path
        name: id
        required: true
        type: string
      - description: Member payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.AddGroupMemberRequest'
      produces:
      - application/json
      responses:
        "201":
          description: message
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: invalid id or payload
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Add a member to a group
      tags:
      - group-members
  /groups/{id}/members/{memberID}:
    delete:
      description: Removes the specified member from the group.
      parameters:
      - description: Group UUID
        in: path
        name: id
        required: true
        type: string
      - description: Member UUID
        in: path
        name: memberID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: message
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: invalid id
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Remove a member from a group
      tags:
      - group-members
  /me:
    get:
      description: Returns the profile of the authenticated user. UUID is extracted
        from Firebase token claims — no extra DB lookup.
      produces:
      - application/json
      responses:
        "200":
          description: current user
          schema:
            additionalProperties: true
            type: object
        "401":
          description: unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: user not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - auth
  /users:
    get:
      description: Returns all non-deleted users.
      produces:
      - application/json
      responses:
        "200":
          description: list of users
          schema:
            additionalProperties: true
            type: object
        "500":
          description: internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List all users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Creates a user record directly. For public registration use POST
        /auth/signup instead.
      parameters:
      - description: Create user payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: created user
          schema:
            additionalProperties: true
            type: object
        "400":
          description: validation error
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a user
      tags:
      - users
  /users/{id}:
    delete:
      description: Sets is_deleted = true. Does not remove the row.
      parameters:
      - description: User UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: message
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: invalid id
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Soft-delete a user
      tags:
      - users
    get:
      description: Returns a single user by UUID.
      parameters:
      - description: User UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: user
          schema:
            additionalProperties: true
            type: object
        "400":
          description: invalid id
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: user not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get user by ID
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Updates mutable fields — username, full name, phone, currency preference,
        password.
      parameters:
      - description: User UUID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: updated user
          schema:
            additionalProperties: true
            type: object
        "400":
          description: invalid id or payload
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a user
      tags:
      - users
  /users/groups:
    get:
      description: Returns all groups the authenticated user is a member of.
      produces:
      - application/json
      responses:
        "200":
          description: groups
          schema:
            items:
              $ref: '#/definitions/models.Group'
            type: array
        "401":
          description: unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get groups for current user
      tags:
      - groups
securityDefinitions:
  BearerAuth:
    description: 'Enter: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"